Move add_related_media_to_objects to base_media_test
