+++ firstvoices/backend/tests/test_apis/test_audio_api.py
@@ -20,6 +20,7 @@ class TestAudioEndpoint(BaseMediaApiTest):
     sample_filetype = "audio/mpeg"
     model = Audio
     model_factory = factories.AudioFactory
+    related_key = "related_audio"
     content_type_json = "application/json"
 
     def create_original_instance_for_patch(self, site):
@@ -271,46 +272,3 @@ class TestAudioEndpoint(BaseMediaApiTest):
         )
         response_data = json.loads(response.content)
         assert response_data["speakers"] == []
-
-    def add_related_media_to_objects(self, visibility=Visibility.PUBLIC):
-        if visibility == Visibility.TEAM:
-            site = self.create_site_with_non_member(Visibility.PUBLIC)
-        else:
-            site = self.create_site_with_app_admin(Visibility.PUBLIC)
-
-        instance = self.create_minimal_instance(site, visibility)
-
-        character = factories.CharacterFactory(site=site, title="a", sort_order=1)
-        character.related_audio.add(instance)
-
-        dict_entry = factories.DictionaryEntryFactory(site=site, visibility=visibility)
-        dict_entry.related_audio.add(instance)
-
-        song = factories.SongFactory(site=site, visibility=visibility)
-        song.related_audio.add(instance)
-
-        story_1 = factories.StoryFactory(site=site, visibility=visibility)
-        story_1.related_audio.add(instance)
-
-        story_page_1 = factories.StoryPageFactory(
-            site=site, story=story_1, visibility=visibility
-        )
-        story_page_1.related_audio.add(instance)
-
-        story_2 = factories.StoryFactory(site=site, visibility=visibility)
-        story_page_2 = factories.StoryPageFactory(
-            site=site, story=story_2, visibility=visibility
-        )
-        story_page_2.related_audio.add(instance)
-
-        total = 5
-
-        return {
-            "site": site,
-            "media_instance": instance,
-            "character": character,
-            "dict_entry": dict_entry,
-            "song": song,
-            "stories": [story_1, story_2],
-            "total": total,
-        }
