+++ firstvoices/backend/tests/test_apis/test_images_api.py
@@ -20,6 +20,7 @@ class TestImagesEndpoint(BaseVisualMediaAPITest):
     sample_filetype = "image/jpg"
     model = Image
     model_factory = factories.ImageFactory
+    related_key = "related_images"
 
     def get_expected_response(self, instance, site, detail_view=False):
         return self.get_expected_image_data(instance, detail_view)
@@ -36,48 +37,6 @@ class TestImagesEndpoint(BaseVisualMediaAPITest):
                 expected.pop(ignored_field)
 
         assert actual_response == expected
-
-    def add_related_media_to_objects(self, visibility=Visibility.PUBLIC):
-        if visibility == Visibility.TEAM:
-            site = self.create_site_with_non_member(Visibility.PUBLIC)
-        else:
-            site = self.create_site_with_app_admin(Visibility.PUBLIC)
-        instance = self.create_minimal_instance(site, visibility=Visibility.PUBLIC)
-
-        character = factories.CharacterFactory(site=site, title="a", sort_order=1)
-        character.related_images.add(instance)
-
-        dict_entry = factories.DictionaryEntryFactory(site=site, visibility=visibility)
-        dict_entry.related_images.add(instance)
-
-        song = factories.SongFactory(site=site, visibility=visibility)
-        song.related_images.add(instance)
-
-        story_1 = factories.StoryFactory(site=site, visibility=visibility)
-        story_1.related_images.add(instance)
-
-        story_page_1 = factories.StoryPageFactory(
-            site=site, story=story_1, visibility=visibility
-        )
-        story_page_1.related_images.add(instance)
-
-        story_2 = factories.StoryFactory(site=site, visibility=visibility)
-        story_page_2 = factories.StoryPageFactory(
-            site=site, story=story_2, visibility=visibility
-        )
-        story_page_2.related_images.add(instance)
-
-        total = 5
-
-        return {
-            "site": site,
-            "media_instance": instance,
-            "character": character,
-            "dict_entry": dict_entry,
-            "song": song,
-            "stories": [story_1, story_2],
-            "total": total,
-        }
 
     @pytest.mark.django_db
     def test_usages_field_extra_fields(self):
