+++ firstvoices/backend/tests/test_apis/test_videos_api.py
@@ -20,6 +20,7 @@ class TestVideosEndpoint(BaseVisualMediaAPITest):
     sample_filetype = "video/mp4"
     model = Video
     model_factory = factories.VideoFactory
+    related_key = "related_videos"
 
     def get_expected_response(self, instance, site, detail_view=False):
         return self.get_expected_video_data(instance, detail_view)
@@ -37,48 +38,6 @@ class TestVideosEndpoint(BaseVisualMediaAPITest):
                 expected.pop(ignored_field)
 
         assert actual_response == expected
-
-    def add_related_media_to_objects(self, visibility=Visibility.PUBLIC):
-        if visibility == Visibility.TEAM:
-            site = self.create_site_with_non_member(Visibility.PUBLIC)
-        else:
-            site = self.create_site_with_app_admin(Visibility.PUBLIC)
-        instance = self.create_minimal_instance(site, visibility=Visibility.PUBLIC)
-
-        character = factories.CharacterFactory(site=site, title="a", sort_order=1)
-        character.related_videos.add(instance)
-
-        dict_entry = factories.DictionaryEntryFactory(site=site, visibility=visibility)
-        dict_entry.related_videos.add(instance)
-
-        song = factories.SongFactory(site=site, visibility=visibility)
-        song.related_videos.add(instance)
-
-        story_1 = factories.StoryFactory(site=site, visibility=visibility)
-        story_1.related_videos.add(instance)
-
-        story_page_1 = factories.StoryPageFactory(
-            site=site, story=story_1, visibility=visibility
-        )
-        story_page_1.related_videos.add(instance)
-
-        story_2 = factories.StoryFactory(site=site, visibility=visibility)
-        story_page_2 = factories.StoryPageFactory(
-            site=site, story=story_2, visibility=visibility
-        )
-        story_page_2.related_videos.add(instance)
-
-        total = 5
-
-        return {
-            "site": site,
-            "media_instance": instance,
-            "character": character,
-            "dict_entry": dict_entry,
-            "song": song,
-            "stories": [story_1, story_2],
-            "total": total,
-        }
 
     @pytest.mark.django_db
     def test_usages_field_extra_fields(self):
